refactor: 124-rename the create-personal-info-section.ts file to personal-info.ts
