+++ src/components/profile/layout.ts
@@ -1,7 +1,7 @@
 import { PROFILE_CLASSES } from '../../pages/profile/constants';
 import { HEADER2 } from '../../shared/styles';
 import { createDiv, createH2 } from '../../utils/create-elements/create-tags';
-import { createPersonalInfoSection } from './personal-info-section/create-personal-info-section';
+import { createPersonalInfoSection } from './personal-info-section/personal-info';
 
 export async function profileLayout(): Promise<HTMLDivElement | void> {
   const title = createH2({ classes: HEADER2.general, text: 'Personal Profile' });
